refactor: move mocked IAM client to testing utils
This moves the mocked IAM client of the service-user test to our testing utils
